+++ api/general/routes.py
@@ -2,13 +2,11 @@ from fastapi import APIRouter
 
 from api.general.index import router as index_router
 from api.general.auth import router as auth_router
-from api.general.auth_db import router as auth_db_router
 
 routers = APIRouter()
 router_list = [
     index_router,
-    auth_router,
-    auth_db_router
+    auth_router
 ]
 
 for router in router_list:
